Update README.md
All of the information sent through any and all forms of mediums of communication by the SUPERIOR CLAIMANT are private.  Including but not limited to, any attachment(s), and are limited for the sole use of the intended recipient and may contain Privileged and/or Confidential Information.  If you are not the addressee, or are not authorized to receive for the intended addressee, you are hereby notified that you may not use, copy, distribute, or disclose to anyone this message or the information contained herein. If you have received this message in error, immediately advise the sender by reply email and destroy this message.  ECPA 18 USCA 2510.Gramm-Leach Bliley Act,15 USC Subchapter I, Sec.6801-6809, Title 18 chpt.121 §§ 2701,2702,2709(c)(e)Title 18 chpt.§ 1505/1519/2239/chpt.109 § 2331; All Rights Reserved BY CFO; UNIVERSAL TEMPLE of MA'AT-YHWH TM